// mssql-js/lib/datatypes/enums.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

export enum JsSqlDataTypes {
  Void = 31,
  Image = 34,
  Text = 35,
  UniqueIdentifier = 36,
  VarChar = 39,
  Date = 40,
  Time = 41,
  DateTime2 = 42,
  DateTimeOffset = 43,
  Char = 47,
  TinyInt = 48,
  Bit = 50,
  SmallInt = 52,
  Decimal = 55,
  Int = 56,
  SmallDateTime = 58,
  Real = 59,
  Money = 60,
  DateTime = 61,
  Float = 62,
  Numeric = 63,
  SsVariant = 98,
  NText = 99,
  FltN = 109,
  SmallMoney = 122,
  BigInt = 127,
  BigVarBinary = 165,
  BigVarChar = 167,
  BigBinary = 173,
  BigChar = 175,
  NVarChar = 231,
  NChar = 239,
  Udt = 240,
  Xml = 241,
  Json = 244,
}